% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttapi2-package.R
\docType{package}
\name{ttapi2-package}
\alias{ttapi2}
\alias{ttapi2-package}
\title{ttapi2: TidyTuesday Dataset API}
\description{
Query up-to-date metadata about TidyTuesday datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/ttapi2}
  \item \url{https://r4ds.github.io/ttapi2/}
  \item Report bugs at \url{https://github.com/r4ds/ttapi2/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
